requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
python-dotenv==1.0.0
mysqlclient==2.2.0
qrcode==7.4.2
PyPDF2==3.0.1
reportlab==4.0.7
python-dateutil==2.8.2